Move macros to task SDK execution_time module
