+++ airflow-core/tests/unit/plugins/test_plugins_manager.py
@@ -229,17 +229,17 @@ class TestPluginsManager:
         Tests whether macros that originate from plugins are being registered correctly.
         """
         from airflow.plugins_manager import integrate_macros_plugins
-        from airflow.sdk.definitions import macros
+        from airflow.sdk.execution_time import macros
 
         def cleanup_macros():
             """Reloads the macros module such that the symbol table is reset after the test."""
             # We're explicitly deleting the module from sys.modules and importing it again
             # using import_module() as opposed to using importlib.reload() because the latter
-            # does not undo the changes to the airflow.sdk.definitions.macros module that are being caused by
+            # does not undo the changes to the airflow.sdk.execution_time.macros module that are being caused by
             # invoking integrate_macros_plugins()
 
-            del sys.modules["airflow.sdk.definitions.macros"]
-            importlib.import_module("airflow.sdk.definitions.macros")
+            del sys.modules["airflow.sdk.execution_time.macros"]
+            importlib.import_module("airflow.sdk.execution_time.macros")
 
         request.addfinalizer(cleanup_macros)
 
@@ -254,7 +254,7 @@ class TestPluginsManager:
             # Ensure the macros for the plugin have been integrated.
             integrate_macros_plugins()
             # Test whether the modules have been created as expected.
-            plugin_macros = importlib.import_module(f"airflow.sdk.definitions.macros.{MacroPlugin.name}")
+            plugin_macros = importlib.import_module(f"airflow.sdk.execution_time.macros.{MacroPlugin.name}")
             for macro in MacroPlugin.macros:
                 # Verify that the macros added by the plugin are being set correctly
                 # on the plugin's macro module.
